+++ grails-app/controllers/com/headbangers/mordheim/MaverickController.groovy
@@ -143,7 +143,7 @@ class MaverickController {
             }
 
             ByteArrayOutputStream out = new ByteArrayOutputStream()
-            ImageHelper.resize(f.bytes, out, new Integer(55), new Integer(55))
+            ImageHelper.resize(f.bytes, out, new Integer(135), new Integer(135))
             maverickInstance.photo = out.toByteArray()
 
         } else {
