+++ grails-app/controllers/com/headbangers/mordheim/WrenchmenController.groovy
@@ -201,7 +201,7 @@ class WrenchmenController {
             }
 
             ByteArrayOutputStream out = new ByteArrayOutputStream()
-            ImageHelper.resize(f.bytes, out, new Integer(55), new Integer(55))
+            ImageHelper.resize(f.bytes, out, new Integer(135), new Integer(135))
             wrenchmenInstance.photo = out.toByteArray()
 
         } else {
